Added support for SDO tools generic class.
These classes can be used to fetch and install all tools
delivered by SDO, such as BIOS, XDC, ...

It does not rely internally on dirac fetcher anymore and /data/omapts.

The class sdotools should be used to install SDO tools that come with an
installer. The class sdotools-tar should be used for tools delivered as
tar archives.