From: Nicolas Dechesne <n-dechesne@ti.com> Date: Thu, 3 Sep 2009 00:00:03 +0200 Subject: [PATCH] Added support for SRCREV in ccasefetch class.
It allows using 'generic' recipe for clearcase based components with names such
as mycomponent_cc.bb. That way the recipe does not change from one release to
another. It is similar to the xxx_git.bb which is built in bitbake.

An example of baseimage_cc.bb is:

PRIORITY = "optional"
DESCRIPTION = "Texas Instruments Baseimage."
LICENSE = "LGPL"
PR = "r0"

inherit ccasefetch

PV = "4.0+cc+${SRCREV}"

CCASE_SPEC = "%\
   element /vobs/wtbu/OMAPSW_DSP/system/baseimage/... ${SRCREV}%\
   element * /main/LATEST%"

CCASE_PATHFETCH = "/vobs/wtbu/OMAPSW_DSP/system"
CCASE_PATHCOMPONENT = "OMAPSW_DSP"
CCASE_PATHCOMPONENTS = "2"

...

In this example, you can see that the PV is using the SRCREV variable so that a
different tar ball is created if we update the revision. If you look at the
config spec you can see that the SRCREV is in turn the component label, it is
set in the versions.inc file, such as:

SRCREV_pn-baseimage ?= "ACT_BASEIMAGE-4430-0.6"

And the variable PREFERRED_VERSION_baseimage is not set anymore. SRCREV does
not need to be a clearcase label, it can be anything meaningful from which the
clearcase label can be derived. So we can even do something like if SRCREV
contains ACT_ then SRCREV is the label, other SRCREV is just a version number,
and the label is MY_COMPONENT_REL_SRCREV...

For this example, the tar ball generated is:
baseimage-4.0+cc+ACT_BASEIMAGE-4430-0.6.tar.gz, and the WORKING_DIR is
\build\tmp\work\armv7a-none-linux-gnueabi\baseimage-4.0+cc+ACT_BASEIMAGE-4430-0.6-r0,
as expected.