build gstreamer packages direct from git trees without patches
It is too difficult for customers who build outside of poky to deal with
figuring out which patches to apply, so it is better to handle it like
other components and pull directly from a git tree without patches in
the recipe.

We'll need to move these trees to dev.omapzoom.org, but for now
gitorious will do.  Also, recipes should be updated to use labels rather
than commit-ids.